improve files hierarchy
