+++ src/RerollKnight/Assets/Code/GameStates/StateMachine/States/Prepare/Initialize/InitializeGameplayState.cs
@@ -1,6 +1,4 @@
 using Code.Component;
-using Entitas;
-using Entitas.Generic;
 
 namespace Code
 {
@@ -24,24 +22,6 @@ namespace Code
 				// TODO: is it the best state?
 				Add<ToStateWhenAllReady<TurnEndedGameplayState>>();
 			}
-		}
-	}
-
-	public sealed class PutPlayerFirstSystem : IInitializeSystem
-	{
-		private readonly TurnsQueue _turnsQueue;
-		private readonly IGroup<Entity<GameScope>> _players;
-
-		public PutPlayerFirstSystem(Contexts contexts, TurnsQueue turnsQueue)
-		{
-			_turnsQueue = turnsQueue;
-			_players = contexts.GetGroup(ScopeMatcher<GameScope>.Get<Player>());
-		}
-
-		public void Initialize()
-		{
-			foreach (var player in _players)
-				_turnsQueue.PutFirst(player);
 		}
 	}
 }
