+++ src/RerollKnight/Assets/Code/GameStates/StateMachine/States/Prepare/Initialize/PutPlayerFirstSystem.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 87d03c3418224d318b6274db73fa8df2
+timeCreated: 1700418564
