+++ src/RerollKnight/Assets/Code/GameStates/StateMachine/States/Prepare/Load/LoadAssetsGameplayState.cs.meta
